HouseDetailDataSource configuring

+++ View/HouseDetailsTableViewCell.swift
@@ -15,7 +15,7 @@ class HouseDetailsTableViewCell: UITableViewCell {
     @IBOutlet weak var housePriceTextLabel: UILabel!
     @IBOutlet weak var houseAreaTextLabel: UILabel!
     
-    func configure(with adress: String, price: Double, area: Int, rooms: Int) {
+    func configure(with adress: String, price: Int, area: Int, rooms: Int) {
         
         housePriceTextLabel.text = "\(price) €/month"
         houseAreaTextLabel.text = "\(area)^2, \(rooms) rooms "
